+++ library/src/scripts/layout/Backgrounds.tsx
@@ -8,6 +8,7 @@ import React, { ReactNode, useContext, useEffect, useState } from "react";
 import { fullBackgroundClasses, bodyCSS } from "@library/layout/bodyStyles";
 import { useHistory } from "react-router";
 import { LoadStatus } from "@library/@types/api/core";
+import "@library/theming/reset";
 
 interface IProps {
     isHomePage?: boolean;
