+++ library/src/scripts/theming/ThemeProvider.tsx
@@ -16,8 +16,6 @@ import { useSelector } from "react-redux";
 import { loadThemeFonts } from "./loadThemeFonts";
 import { Backgrounds, BackgroundsProvider } from "@library/layout/Backgrounds";
 import { BrowserRouter } from "react-router-dom";
-import {ErrorPage} from "@library/errorPages/ErrorComponent";
-import {DefaultError} from "@library/errorPages/ErrorMessagePage";
 
 interface IProps {
     children: React.ReactNode;
@@ -80,13 +78,8 @@ export const ThemeProvider: React.FC<IProps> = (props: IProps) => {
         return <Loader />;
     }
 
-    if (assets.status === LoadStatus.SUCCESS) {
-        console.log('',assets);
-        return(
-            <BrowserRouter>
-                <ErrorPage defaultError={DefaultError.NOT_FOUND} code={404} />
-            </BrowserRouter>
-        );
+    if (assets.status === LoadStatus.ERROR) {
+        return <>{props.errorComponent}</>;
     }
 
     if (!assets.data) {
@@ -95,7 +88,9 @@ export const ThemeProvider: React.FC<IProps> = (props: IProps) => {
 
     // Apply kludged input text styling everywhere.
     inputClasses().applyInputCSSRules();
-    console.log(assets.status);
+
+
+    console.log("rendering backgrounds");
     return (
         <>
             <BrowserRouter>
